Change 'index' arg for pdfi_dict_next() and _first()
It was a "void *", which just seems like asking for trouble.

Changed it to a "uint64_t *", which is the type of the corresponding
values in the pdf_dict struct.  Fixed all the callers to use that
type.

+++ pdf/pdf_colour.c
@@ -1821,7 +1821,7 @@ static int pdfi_create_DeviceN(pdf_context *ctx, pdf_array *color_array, int ind
             goto pdfi_devicen_error;
 
         if (Colorants != NULL && pdfi_dict_entries(Colorants) != 0) {
-            int ix = 0;
+            uint64_t ix = 0;
             pdf_obj *Colorant = NULL, *Space = NULL;
             char *colorant_name;
             gs_color_space *colorant_space;
